@@ -1,41 +1,42 @@
 {
-  "name": "tailwindui-template",
+  "name": "portfolio",
+  "author": "Ryan Freeman",
   "version": "0.1.0",
   "private": true,
   "scripts": {
     "dev": "next dev",
-    "build": "next build && next-sitemap",
+    "build": "next build",
     "start": "next start",
     "lint": "next lint"
   },
-  "browserslist": "defaults, not ie <= 11",
   "dependencies": {
-    "@headlessui/react": "^1.7.0",
+    "@headlessui/react": "^1.7.7",
     "@mapbox/rehype-prism": "^0.8.0",
-    "@mdx-js/loader": "^2.1.2",
-    "@next/mdx": "^12.2.3",
-    "@tailwindcss/typography": "^0.5.4",
-    "autoprefixer": "^10.4.7",
-    "clsx": "^1.2.0",
-    "fast-glob": "^3.2.11",
+    "@next/mdx": "^13.1.1",
+    "@tailwindcss/line-clamp": "^0.4.2",
+    "@tailwindcss/typography": "^0.5.8",
+    "@types/mdx": "^2.0.3",
+    "@types/node": "18.11.18",
+    "@types/react": "18.0.26",
+    "@types/react-dom": "18.0.10",
+    "autoprefixer": "^10.4.13",
+    "clsx": "^1.2.1",
+    "eslint": "8.31.0",
+    "eslint-config-next": "13.1.1",
+    "fast-glob": "^3.2.12",
     "feed": "^4.2.2",
     "focus-visible": "^5.2.0",
-    "motion": "^10.15.3",
-    "next": "^12.3.0",
-    "next-sitemap": "^3.1.44",
+    "motion": "^10.15.5",
+    "next": "13.1.2",
     "node-fetch": "^3.3.0",
-    "postcss-focus-visible": "^6.0.4",
+    "postcss": "^8.4.21",
+    "postcss-focus-visible": "^7.1.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "remark-gfm": "^3.0.1",
+    "sharp": "^0.31.3",
     "swr": "^2.0.0",
-    "tailwindcss": "^3.1.4"
-  },
-  "devDependencies": {
-    "@tailwindcss/line-clamp": "^0.4.2",
-    "eslint": "8.19.0",
-    "eslint-config-next": "12.2.5",
-    "prettier": "^2.7.1",
-    "prettier-plugin-tailwindcss": "^0.1.11"
+    "tailwindcss": "^3.2.4",
+    "typescript": "4.9.4"
   }
 }
