@@ -1,30 +1,27 @@
 import Head from 'next/head'
-import {useRouter} from 'next/router'
-
-import {Container} from '@/components/Container'
+import {usePathname} from 'next/navigation'
+import {Container} from './Container'
 import {formatDate} from '@/lib/formatDate'
-import {Prose} from '@/components/Prose'
-
-function ArrowLeftIcon(props) {
-    return (
-        <svg viewBox="0 0 16 16" fill="none" aria-hidden="true" {...props}>
-            <path
-                d="M7.25 11.25 3.75 8m0 0 3.5-3.25M3.75 8h8.5"
-                strokeWidth="1.5"
-                strokeLinecap="round"
-                strokeLinejoin="round"
-            />
-        </svg>
-    )
-}
+import {Prose} from './Prose'
+import {ReactNode} from 'react'
 
 export function ArticleLayout({
                                   children,
-                                  meta,
                                   isRssFeed = false,
-                                  previousPathname,
-                              }) {
-    let router = useRouter()
+                                  title,
+                                  description,
+                                  ogImage,
+                                  date
+                              }:
+                                  {
+                                      children?: ReactNode,
+                                      isRssFeed: boolean,
+                                      title: string,
+                                      description: string,
+                                      ogImage: string,
+                                      date: string
+                                  }) {
+    const pathname = usePathname()
 
     if (isRssFeed) {
         return children
@@ -33,11 +30,11 @@ export function ArticleLayout({
     return (
         <>
             <Head>
-                <title>{`${meta.title} - Ryan Freeman`}</title>
-                <meta name="description" content={meta.description}/>
+                <title>{`${title} - Ryan Freeman`}</title>
+                <meta name="description" content={description}/>
                 <meta
                     property="og:url"
-                    content={`${process.env.NEXT_PUBLIC_SITE_URL}${router.pathname}`}
+                    content={`${process.env.NEXT_PUBLIC_SITE_URL}${pathname}`}
                 />
                 <meta
                     property="og:type"
@@ -45,17 +42,17 @@ export function ArticleLayout({
                 />
                 <meta
                     property="og:title"
-                    content={meta.title}
+                    content={title}
                 />
                 <meta
                     property="og:description"
-                    content={meta.description}
+                    content={description}
                 />
-                {meta.ogImage &&
+                {ogImage &&
                     <>
                         <meta
                             property="og:image"
-                            content={meta.ogImage}
+                            content={ogImage}
                         />
                         <meta
                             name="twitter:card"
@@ -63,7 +60,7 @@ export function ArticleLayout({
                         />
                         <meta
                             name="twitter:image"
-                            content={meta.ogImage}
+                            content={ogImage}
                         />
                     </>
                 }
@@ -73,41 +70,30 @@ export function ArticleLayout({
                 />
                 <meta
                     property="twitter:url"
-                    content={`${process.env.NEXT_PUBLIC_SITE_URL}${router.pathname}`}
+                    content={`${process.env.NEXT_PUBLIC_SITE_URL}${pathname}`}
                 />
                 <meta
                     name="twitter:title"
-                    content={meta.title}
+                    content={title}
                 />
                 <meta
                     name="twitter:description"
-                    content={meta.description}
+                    content={description}
                 />
             </Head>
             <Container className="mt-16 lg:mt-32">
                 <div className="xl:relative">
                     <div className="mx-auto max-w-2xl">
-                        {previousPathname && (
-                            <button
-                                type="button"
-                                onClick={() => router.back()}
-                                aria-label="Go back to articles"
-                                className="group mb-8 flex h-10 w-10 items-center justify-center rounded-full bg-white shadow-md shadow-zinc-800/5 ring-1 ring-zinc-900/5 transition dark:border dark:border-zinc-700/50 dark:bg-zinc-800 dark:ring-0 dark:ring-white/10 dark:hover:border-zinc-700 dark:hover:ring-white/20 lg:absolute lg:-left-5 lg:mb-0 lg:-mt-2 xl:-top-1.5 xl:left-0 xl:mt-0"
-                            >
-                                <ArrowLeftIcon
-                                    className="h-4 w-4 stroke-zinc-500 transition group-hover:stroke-zinc-700 dark:stroke-zinc-500 dark:group-hover:stroke-zinc-400"/>
-                            </button>
-                        )}
                         <article>
                             <header className="flex flex-col">
                                 <h1 className="mt-6 text-4xl font-bold tracking-tight text-zinc-800 dark:text-zinc-100 sm:text-5xl">
-                                    {meta.title}
+                                    {title}
                                 </h1>
                                 <time
-                                    dateTime={meta.date}
+                                    dateTime={date}
                                     className="order-first flex items-center text-base text-zinc-500 dark:text-zinc-400"
                                 >
-                                    <span>{formatDate(meta.date)}</span>
+                                    <span>{formatDate(date)}</span>
                                 </time>
                             </header>
                             <Prose className="mt-8">{children}</Prose>
