@@ -1,9 +1,9 @@
 module.exports = {
-  plugins: {
-    tailwindcss: {},
-    'postcss-focus-visible': {
-      replaceWith: '[data-focus-visible-added]',
-    },
-    autoprefixer: {},
-  },
+    plugins: {
+        tailwindcss: {},
+        'postcss-focus-visible': {
+            replaceWith: '[data-focus-visible-added]',
+        },
+        autoprefixer: {},
+    }
 }
