@@ -1,6 +1,6 @@
 /** @type {import('tailwindcss').Config} */
 module.exports = {
-    content: ['./src/**/*.{js,jsx}'],
+    content: ['./src/**/*.{js,jsx,tsx}'],
     darkMode: 'class',
     plugins: [
         require('@tailwindcss/typography'),
